Add question order and index handling and answer push api

+++ main.go
@@ -1,5 +1,6 @@
 package main
 import (
+    "os"
     "math/rand"
     "strconv"
     "log"
@@ -26,6 +27,10 @@ type User struct {
     Password string `json:"password"`
     Token string `json:"token"`
     GridNumbers []int `json:"gridnumbers"`
+    QuestionOrder []int `json:"questionorder"`
+    QuestionIndex int `json:"questionindex"`
+    AnswerText string `json:"answertext"`
+    AnswerBase64Str string `json:"answerbase64str"`
 }
 
 type Question struct {
@@ -55,12 +60,14 @@ func destroyMongoClient(client *mongo.Client) {
     fmt.Println("[*] Client Destruction Success!")
 }
 
-func createDatabaseCollection (client *mongo.Client) *mongo.Collection{
+func createDatabaseCollection (client *mongo.Client, to_clear_collection bool) *mongo.Collection{
     collection := client.Database(databaseName).Collection("users")
     // Clear all users
-    _, err := collection.DeleteMany(context.TODO(), bson.D{{}})
-    if err != nil {
-        log.Fatal(err)
+    if to_clear_collection {
+        _, err := collection.DeleteMany(context.TODO(), bson.D{{}})
+        if err != nil {
+            log.Fatal(err)
+        }
     }
     return collection
 }
@@ -115,7 +122,9 @@ func check_authentication(collection *mongo.Collection) gin.HandlerFunc{
         filter := bson.D{{"password", user.Token}}
         err := collection.FindOne(context.TODO(), filter).Decode(&user)
         if err != nil {
-            log.Fatal(err)
+            log.Printf("[*] Did not find any team match password: %s", user.Token)
+            c.AbortWithStatus(http.StatusNotAcceptable)
+            return
         }
         log.Printf("[*] Get account: %s", user.Account)
         c.Set("account", user.Account)
@@ -124,30 +133,75 @@ func check_authentication(collection *mongo.Collection) gin.HandlerFunc{
     }
 }
 
-func (user *User) saveGridNumbers(collection *mongo.Collection) error{
-    fmt.Println("haha")
+func (user *User) saveGridNumbersAndIntialize(collection *mongo.Collection) bool{
+    var question_order []int
+    question_index := 0
+    for i:= 0; i < 9; i++ {
+        question_order = append(question_order, i)
+    }
+    // shuffle question order
+    rand.Shuffle(len(question_order), func(i, j int){
+        question_order[i], question_order[j] = question_order[j], question_order[i]
+    })
+    // check if user's grid numbers is valid
+    log.Printf("User %s submitted grid numbers are: %v", user.GridNumbers)
+    if len(user.GridNumbers) != 9 {
+        log.Printf("Grid numbers' length should be 9!")
+        return false
+    }
     _, err := collection.UpdateOne(context.TODO(),
         bson.D{
             { "account", user.Account },
         }, bson.D{
             {"$set", bson.D{
                         {"gridnumbers", user.GridNumbers},
+                        {"questionorder", question_order},
+                        {"questionindex", question_index},
                     },
             },
         },
     )
-    return err
+    // Fail
+    if err != nil {
+        log.Printf("Something went wrong when user %s try to submit his grid numbers, his grid numbers are: %v", user.Account, user.GridNumbers)
+        return false
+    }
+    // Success
+    user.QuestionOrder = question_order
+    user.QuestionIndex = question_index
+    return true
 }
 
-func (user *User) getGridNumbersByAccount(collection *mongo.Collection){
+func (user *User) getGridNumbersByAccount(collection *mongo.Collection) bool{
+    log.Printf("[*] User %s try to get grid numbers", user.Account)
+    err := collection.FindOne(context.TODO(),
+            bson.D{{"account", user.Account}},
+        ).Decode(user)
+
+    if err != nil {
+        // Team not found or database error
+        return false
+    }
+
+    if user.GridNumbers == nil {
+        log.Printf("[*] User %s's grid numbers have not been submitted yet", user.Account)
+        return false
+    }
+    log.Printf("[*] User %s's grid numbers are: %v", user.Account, user.GridNumbers)
+    return true
+}
+
+func (user *User) saveAnswer(){
+    // Find this user by token(password) and update its answer
 }
 
 func main(){
+    to_clear_collection := os.Args[1]
     //
     rand.Seed(115813)
     // 
     client := createMongoClient()
-    collection := createDatabaseCollection(client)
+    collection := createDatabaseCollection(client, to_clear_collection == "true")
     initialize_users(collection, 10)
 
     router := gin.Default()
@@ -159,10 +213,21 @@ func main(){
     // private part (authentication required)
     {
         private_router.POST("/get_gridnumbers", func(c *gin.Context){
-            // account, _ := c.MustGet("account").(string)
+            var user User
+            account, _ := c.MustGet("account").(string)
+            user.Account = account
+            found := user.getGridNumbersByAccount(collection)
 
+            if !found {
+                c.AbortWithStatus(http.StatusNotFound)
+                return
+            }
+
+            // Return grid numbers to the front end client
             c.JSON(http.StatusOK, gin.H{
-                "gridNumbers": []int{1,2,3,4,5,6,7,8,9},
+                "gridNumbers": user.GridNumbers,
+                "questionOrder": user.QuestionOrder,
+                "questionIndex": user.QuestionIndex,
             })
             return
         })
@@ -174,20 +239,38 @@ func main(){
             if c.ShouldBindBodyWith(&user, binding.JSON) == nil{
                 log.Printf("Get %s's gridNumbers: %v\n", account, user.GridNumbers)
                 // Save grid numbers to database
-                err := user.saveGridNumbers(collection)
+                success := user.saveGridNumbersAndIntialize(collection)
                 // if error
-                if err != nil {
+                if !success {
                     c.String(http.StatusNotFound, notfound)
                     return
                 }
 
                 c.JSON(http.StatusOK, gin.H{
                     "gridNumbers": user.GridNumbers,
+                    "questionOrder": user.QuestionOrder,
+                    "questionIndex": user.QuestionIndex,
                 })
             }else{
                 c.String(http.StatusNotFound, "Page Not Found")
             }
             return
+        })
+
+        private_router.POST("/push_answer", func (c *gin.Context){
+            // account, _ := c.MustGet("account").(string)
+            var user User
+            if c.ShouldBindBodyWith(&user, binding.JSON) == nil {
+                // TODO: save answer text and image base64 string to database
+                c.String(http.StatusOK, "Answer Received")
+            }else{
+                c.String(http.StatusNotAcceptable, "Input is weird")
+            }
+        })
+
+        // A user can check what he have answered for this question
+        private_router.GET("/get_previous_answer", func(c *gin.Context){
+            // TODO:
         })
     }
